web-app/plog_page/content_context/deepblue-CMC-chatbot/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
psycopg2-binary

# Vision Model Dependencies
torch
torchvision
transformers
accelerate
pillow
python-multipart
